raiden: transport: move _handle_candidates_callback to WebRTCManager
